move gql.ts file into graphql  folder
