+++ next/components/molecules/sections/general/ArticlesListSection/ArticlesList.tsx
@@ -1,10 +1,10 @@
 import { BlogPostEntity } from '@backend/graphql'
+import { client } from '@backend/graphql/gql'
 import { ArticlesFilter, Card } from '@bratislava/ui-bratislava/ArticlesFilter/ArticlesFilter'
 import { Pagination } from '@bratislava/ui-bratislava/Pagination/Pagination'
 import BlogPostCard from '@components/molecules/presentation/BlogPostCard'
 import { getCategoryColorLocalStyle } from '@utils/colors'
 import { generateImageSizes } from '@utils/generateImageSizes'
-import { client } from '@utils/gql'
 import { getNumericLocalDate } from '@utils/local-date'
 import { useLocale, useTranslations } from 'next-intl'
 import { useEffect, useState } from 'react'
