+++ next/pages/blog/[slug].tsx
@@ -1,9 +1,9 @@
 import { BlogPostEntityFragment, GeneralQuery } from '@backend/graphql'
+import { client } from '@backend/graphql/gql'
 import PageLayout from '@components/layouts/PageLayout'
 import BlogPostPageContent from '@components/pages/blogPostPageContent'
 import { GlobalCategoryColorProvider } from '@utils/colors'
 import { GeneralContextProvider } from '@utils/generalContext'
-import { client } from '@utils/gql'
 import { useTitle } from '@utils/useTitle'
 import { GetStaticPaths, GetStaticProps } from 'next'
 import Head from 'next/head'
