+++ next/pages/index.tsx
@@ -1,9 +1,9 @@
 import type { HomepageContext } from '@backend/fetchers/homepageContextFetcher'
 import { homepageContextFetcher } from '@backend/fetchers/homepageContextFetcher'
 import { GeneralQuery } from '@backend/graphql'
+import { client } from '@backend/graphql/gql'
 import HomepageContent from '@components/pages/HomepageContent'
 import { GeneralContextProvider } from '@utils/generalContext'
-import { client } from '@utils/gql'
 import { HomepageContextProvider } from '@utils/homepageContext'
 import { useTitle } from '@utils/useTitle'
 import { GetStaticProps } from 'next'
