+++ next/pages/vyhladavanie.tsx
@@ -1,8 +1,8 @@
 import { GeneralQuery } from '@backend/graphql'
+import { client } from '@backend/graphql/gql'
 import SearchPageContent from '@components/pages/searchPageContent'
 import { LocalizationsProvider } from '@components/providers/LocalizationsProvider'
 import { GeneralContextProvider } from '@utils/generalContext'
-import { client } from '@utils/gql'
 import { useTitle } from '@utils/useTitle'
 import { GetStaticProps } from 'next'
 import Head from 'next/head'
